// src/api/generated/models/LoginDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { JwtToken } from './JwtToken';
import type { RefreshToken } from './RefreshToken';
export type LoginDto = {
    token?: JwtToken;
    refreshToken?: RefreshToken;
};